requirements.txt
streamlit==1.39.0
sentence-transformers==3.0.1
faiss-cpu
PyPDF2==4.0.1
python-dotenv==1.0.0
groq==0.4.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
huggingface-hub==0.20.0
torch==2.1.1
numpy==1.24.3
scikit-learn==1.3.2